static rules: add processors json config (#10)
The processors config for schema extraction is currently not part of the
defaul ruleset and needs to be added by the libraries. We will need to
access it both from datadog-agent and dd-trace-go so it makes sense to
export it along the default ruleset.

Moving embed to a dedicated file so that we don't keep API code in a txt
template.