fix: possible goroutine leak in limiter.TokenTicker
There is a potential goroutine leak in tests for `limiter.TokenTicker`,
as it cannot be guaranteed that the `stopChan` closing is picked up by
`select` before the closing of the `ticker` channel.

In addition, the closing of the `ticker` channel being picked up first
could possibly result in corruption of the internal state of the ticker,
as the channel would have immediately returned a 0-value (which is
likely long before the latest timestamp); offsetting the `elapsedNs`
count.

This addesses these issues by:
- Draining the `syncChan` upon stopping a `TestTicker` to ensure no send
  to it could result in blocking;
- Moving responsibility of creating the `syncChan` to the caller of
  `start`, as it provides a cleaner API;
- Use the two-value form of channel receives from the ticksChan in order
  to appropriately ignore zero-values read from a closed channel;
- Pass all channels as parameters to the `updateBucket` goroutine, so we
  can `nil` out the fields on the struct when stopped without risking
  to cause the goroutine exit condition to never materialize because the
  instance's `stopChan` is `nil`.