fix: make appsec.DefaultRuleset reutrn the ruleset as-is (#46)
The `DefaultRuleset` function used to re-marshal a JSON document that
was just parsed; as an artifact from times where the processors were
merged into the ruleset separately. This has not been needed since a
long time, as the default ruleset now includes the relevant processors
already.

This removes inefficiencies as we no longer have to parse, then format
the data. The embedded object is also carried directly as a `[]byte`
slice instead of a string, to remove the need for a string copy to
happen all the time.

Co-authored-by: Eliott Bouhana <47679741+eliottness@users.noreply.github.com>